refactor

+++ scripts/make_vs_commits.sh
@@ -2,22 +2,48 @@
 
 set -ex
 
+is_in_local () {
+    local branch=${1}
+    local existed_in_local=$(git branch --list ${branch})
+
+    if [[ -z ${existed_in_local} ]]; then
+        echo 0
+    else
+        echo 1
+    fi
+}
+
+is_in_remote() {
+    local branch=${1}
+    local existed_in_remote=$(git ls-remote --heads origin ${branch})
+
+    if [[ -z ${existed_in_remote} ]]; then
+        echo 0
+    else
+        echo 1
+    fi
+}
+
 add_vs_commit () {
-    repo=$1
-    cd ../$1
+    repo=${1}
+    cd ../${repo}
     git status
     git config --local user.email "mszhanyi@users.noreply.github.com"
     git config --local user.name "mszhanyi"
-    git remote add upstream https://github.com/pytorch/$1
+    git remote add upstream https://github.com/pytorch/${repo}
     git remote -v
     git fetch upstream
     git merge upstream/master
     git status
-    git remote set-url origin https://mszhanyi:${pytorch_token}@github.com/mszhanyi/$1.git
+    git remote set-url origin https://mszhanyi:${pytorch_token}@github.com/mszhanyi/${repo}.git
     git push
 
-    git branch -d zhanyi/updatevs
-    git push origin --delete zhanyi/updatevs
+    if [[ -z ${is_in_local "zhanyi/updatevs"} ]]; then
+        git branch -d zhanyi/updatevs
+    fi
+    if [[ -z ${is_in_remote "zhanyi/updatevs"} ]]; then
+        git push origin --delete zhanyi/updatevs
+    fi
 
     git checkout -b zhanyi/updatevs
